Fix usage of float constant reported by gcc

+++ src/switchpanel.c
@@ -399,7 +399,7 @@ WSwitchPanel *wInitSwitchPanel(WScreen *scr, WWindow *curwin, Bool class_only)
 	WMGetScaleBaseFromSystemFont(scr->wmscreen, &wmScaleWidth, &wmScaleHeight);
 
 	icon_size = wPreferences.switch_panel_icon_size;
-	icon_tile_size = (short int)(((float)icon_size * (float)1.2) + 0.5);
+	icon_tile_size = (short int)(((float)icon_size * 1.2F) + 0.5F);
 	border_space = WMScaleY(10);
 	label_height = WMScaleY(25);
